jewelry_pos/requirements.txt
acres==0.5.0
aiofiles==24.1.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.9
botocore==1.40.9
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
customtkinter==5.2.2
darkdetect==0.8.0
dj-database-url==3.0.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework-csv==3.0.2
etelemetry==0.3.1
filelock==3.19.1
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
jmespath==1.0.1
looseversion==1.3.0
lxml==6.0.1
MouseInfo==0.1.3
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==10.4.0
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.30
PyAutoGUI==0.9.54
pycparser==2.22
pydot==4.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.26.4
pyparsing==3.2.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.2
pyxnat==1.6.3
qrcode==8.2
rdflib==7.2.1
requests==2.32.5
s3transfer==0.13.1
scipy==1.16.2
selenium==4.35.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.48.0
traits==7.0.2
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.14.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.11.0
wsproto==1.2.0
xlsxwriter==3.2.9
